Better to return an integer in allinsizer, not many brokers will allow to open a float position.

+++ backtrader/sizers/percents_sizer.py
@@ -43,7 +43,7 @@ class PercentSizer(bt.Sizer):
     def _getsizing(self, comminfo, cash, data, isbuy):
         position = self.broker.getposition(data)
         if not position:
-            size = cash / data.close[0] * (self.params.percents / 100)
+            size = int(cash / data.close[0] * (self.params.percents / 100))
         else:
             size = position.size
